Fix test (phenix_regression/ligand_pipeline/tst_fo_minus_fo.py): test was failing because geometry_minimization created a model cif by default, overwriting the restraints cif needed for phenix.refine

+++ mmtbx/command_line/geometry_minimization.py
@@ -422,10 +422,11 @@ class run(object):
       f.close()
 
     cif_fname = self.result_model_fname.replace('.pdb', '.cif')
-    print("  output file name:", cif_fname, file=self.log)
-    r = self.model.model_as_mmcif(output_cs=self.output_crystal_symmetry)
-    with open(cif_fname, 'w') as f:
-      f.write(r)
+    if not os.path.isfile(cif_fname):
+      print("  output file name:", cif_fname, file=self.log)
+      r = self.model.model_as_mmcif(output_cs=self.output_crystal_symmetry)
+      with open(cif_fname, 'w') as f:
+        f.write(r)
 
     if(self.states_collector):
       self.states_collector.write(
